Extract worker/coordinator internal ws handlers
