+++ pkg/coordinator/routes.go
@@ -1,6 +1,8 @@
 package coordinator
 
-import "github.com/giongto35/cloud-game/v2/pkg/cws/api"
+import (
+	"github.com/giongto35/cloud-game/v2/pkg/cws/api"
+)
 
 // workerRoutes adds all worker request routes.
 func (o *Server) workerRoutes(wc *WorkerClient) {
@@ -13,4 +15,22 @@ func (o *Server) workerRoutes(wc *WorkerClient) {
 	wc.Receive(api.GetRoom, wc.handleGetRoom(o))
 	wc.Receive(api.CloseRoom, wc.handleCloseRoom(o))
 	wc.Receive(api.IceCandidate, wc.handleIceCandidate(o))
+}
+
+// useragentRoutes adds all useragent (browser) request routes.
+func (o *Server) useragentRoutes(bc *BrowserClient) {
+	if o == nil {
+		return
+	}
+
+	bc.Receive(api.Heartbeat, bc.handleHeartbeat())
+	bc.Receive(api.InitWebrtc, bc.handleInitWebrtc(o))
+	bc.Receive(api.Answer, bc.handleAnswer(o))
+	bc.Receive(api.IceCandidate, bc.handleIceCandidate(o))
+	bc.Receive(api.GameStart, bc.handleGameStart(o))
+	bc.Receive(api.GameQuit, bc.handleGameQuit(o))
+	bc.Receive(api.GameSave, bc.handleGameSave(o))
+	bc.Receive(api.GameLoad, bc.handleGameLoad(o))
+	bc.Receive(api.GamePlayerSelect, bc.handleGamePlayerSelect(o))
+	bc.Receive(api.GameMultitap, bc.handleGameMultitap(o))
 }
