+++ pkg/worker/routes.go
@@ -10,4 +10,15 @@ func (h *Handler) routes() {
 	h.oClient.Receive(api.ServerId, h.handleServerId())
 	h.oClient.Receive(api.SignalReady, h.handleSignalReady())
 	h.oClient.Receive(api.ConfPushRoute, h.handleConfPush())
+	h.oClient.Receive(api.TerminateSession, h.handleTerminateSession())
+	h.oClient.Receive(api.InitWebrtc, h.handleInitWebrtc())
+	h.oClient.Receive(api.Answer, h.handleAnswer())
+	h.oClient.Receive(api.IceCandidate, h.handleIceCandidate())
+
+	h.oClient.Receive(api.GameStart, h.handleGameStart())
+	h.oClient.Receive(api.GameQuit, h.handleGameQuit())
+	h.oClient.Receive(api.GameSave, h.handleGameSave())
+	h.oClient.Receive(api.GameLoad, h.handleGameLoad())
+	h.oClient.Receive(api.GamePlayerSelect, h.handleGamePlayerSelect())
+	h.oClient.Receive(api.GameMultitap, h.handleGameMultitap())
 }
